% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for Metrix package.}
\format{
The data is properly formatted for being used with Metrix functions.
}
\usage{
example_data
}
\description{
Example taxa data from two sites (P1 and P2).
}
\seealso{
\link[metrix]{read_data}, \link[metrix]{metrix_table_template}
}
\keyword{datasets}
