% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for metrix package.}
\format{
The data is properly formated for being used with metrix functions
}
\usage{
example_data
}
\description{
A dataset containing entries from to sites.
}
\keyword{datasets}
